LayoutTests/inspector/runtime/callFunctionOn-emulateUserGesture-hasTransientActivation-expected.txt

Tests for Runtime.callFunctionOn emulateUserGesture option.


== Running test suite: Runtime.callFunctionOn
-- Running test case: Runtime.callFunctionOn.emulateUserGesture.hasTransientActivation.false
No transient activation
-- Running test case: Runtime.callFunctionOn.emulateUserGesture.hasTransientActivation.true
Has transient activation
